Parse angle in styles as double number
Similar to VSFilter. Thanks to Iroha for reporting this issue.

Note: this breaks the ABI!

+++ libass/ass_types.h
@@ -50,7 +50,7 @@ typedef struct ass_style {
     double ScaleX;
     double ScaleY;
     double Spacing;
-    int Angle;
+    double Angle;
     int BorderStyle;
     double Outline;
     double Shadow;
